fixed file structure for transitions and cleaned up
