+++ src/lib/components/MuiTransitions/components/MuiSlide/MuiSlide.js
@@ -1,13 +1,14 @@
 import React, { createElement } from 'react';
 import PropTypes from 'prop-types';
-import { Zoom } from '@material-ui/core';
-import StoryBlok from '../../../utils/Storyblok';
-import MuiGridList from '../../MuiGridList/MuiGridList';
+import { Slide } from '@material-ui/core';
+import StoryBlok from '../../../../utils/Storyblok';
+import MuiGridList from '../../../MuiGridList/MuiGridList';
 
-const MuiZoom = ({
+const MuiSlide = ({
   rootClass,
+  direction,
   content,
-  zoomIn,
+  slideIn,
   timeout,
 }) => {
   const components = {
@@ -18,9 +19,10 @@ const MuiZoom = ({
 
   return (
     <div className={styles.root}>
-      <Zoom
-        in={zoomIn}
+      <Slide
+        in={slideIn}
         timeout={timeout}
+        direction={direction}
       >
         <div>
           {content.map((item, index) => createElement(
@@ -28,14 +30,14 @@ const MuiZoom = ({
             Object.assign(item, { key: index }),
           ))}
         </div>
-      </Zoom>
+      </Slide>
     </div>
   );
 };
 
-export default MuiZoom;
+export default MuiSlide;
 
-MuiZoom.propTypes = {
+MuiSlide.propTypes = {
   /**
    * stroyblok multiselect of css classes
    * Mui Override or extend the styles applied to the component.
@@ -45,7 +47,12 @@ MuiZoom.propTypes = {
    * mui props: bool
    * If true, the component will transition in.
    */
-  zoomIn: PropTypes.bool,
+  slideIn: PropTypes.bool,
+  /**
+   * mui props: 'down' | 'left' | 'right' | 'up'
+   * Direction the child node will enter from.
+   */
+  direction: PropTypes.string,
   /**
    * mui props: number | string
    * The duration for the transition, in milliseconds.
@@ -62,8 +69,9 @@ MuiZoom.propTypes = {
   })).isRequired,
 };
 
-MuiZoom.defaultProps = {
-  zoomIn: true,
+MuiSlide.defaultProps = {
+  slideIn: true,
+  direction: 'down',
   timeout: 'auto',
   rootClass: [],
 };
