+++ src/lib/components/StoryBlokPage/components/Blok/Blok.js
@@ -37,7 +37,6 @@ Blok.propTypes = {
    * Hide the given breakpoint(s).
    * */
   only: PropTypes.arrayOf(PropTypes.string),
-
   /**
    * MuiGrid,
   *  MuiHeroHeader,
